requirements.txt
Django==5.0
djangorestframework==3.15
torch>=2.0.0
transformers>=4.40.0
numpy
gunicorn
langchain
langchain-community
transformers
torch